gemini: Kang walleye offsets for GPU composition
Window animations now have rounded corners, meaning that
the shader that we use is more expensive than it used to be.

Because of this, we need to adjust the SF offsets, to make sure
that we'll have enough time to finish the GPU composition without
dropping frames.

These offsets as similar to blueline's, adjusted slightly because
we were getting false "frame missed" signals if composition
starts too early.

Change-Id: I48fede0f855a8e02c624f08817a0c27c1a9a1b09
Signed-off-by: xNombre <kartapolska@gmail.com>
Signed-off-by: Mohammad Hasan Keramat J <ikeramat80@gmail.com>